migrate jira
